refactor(spec-harness): move harness into tests/, make root project habit-hooks
The spec harness is dev/test tooling, not the main project. Collapse the
specharness package into a single tests/harness.py module (parser, executor,
and CLI runner) and point the root pyproject.toml at habit-hooks itself
(package = false; pytest configured with tests/ on the path). Unit tests run
via `uv run pytest`; specs via `uv run python tests/harness.py`.
