+++ pyproject.toml
@@ -1,19 +1,17 @@
 [project]
-name = "spec-harness"
-version = "0.1.0"
-description = "Test harness that runs the executable *.spec.md docs"
+name = "habit-hooks"
+version = "0.0.0"
+description = "Turn best-practice coding advice into AI habits — structural code-smell coaching for AI agents"
 requires-python = ">=3.11"
-dependencies = []
 
-[project.scripts]
-spec-harness = "specharness.__main__:main"
+[tool.uv]
+# habit-hooks itself ships as plugin/skill files, not a Python package; this
+# pyproject only wires up the dev tooling (the spec test harness under tests/).
+package = false
 
 [dependency-groups]
 dev = ["pytest>=8"]
 
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["specharness"]
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+pythonpath = ["tests"]
