--- specharness/__init__.py
@@ -1,5 +1,0 @@
-"""Test harness for executable ``*.spec.md`` docs."""
-
-from .harness import SpecError, SpecFailure, SpecCase, execute, normalize, parse_spec
-
-__all__ = ["SpecError", "SpecFailure", "SpecCase", "execute", "normalize", "parse_spec"]
