+++ tests/harness.py
@@ -1,8 +1,13 @@
-"""Runs executable specs (``docs/**/*.spec.md``).
+"""Test harness that runs the executable specs (``docs/**/*.spec.md``).
 
+This is dev/test tooling for building habit-hooks, not shipped product code.
 The grammar is defined by ``docs/executable_spec.md`` — that file is the
 contract. Markers are matched by base codepoint; any U+FE0F variation selector
 is ignored.
+
+Run the unit tests with ``uv run pytest``; run the specs themselves with
+``uv run python tests/harness.py [docs/foo.spec.md ...]`` (defaults to every
+``docs/**/*.spec.md``).
 """
 
 from __future__ import annotations
@@ -10,6 +15,8 @@ from __future__ import annotations
 import os
 import re
 import subprocess
+import sys
+import tempfile
 from dataclasses import dataclass, field
 from pathlib import Path
 
@@ -311,3 +318,63 @@ def execute(test: SpecCase, workdir: Path, repo_root: Path) -> None:
                     f"--- actual ---\n{normalize(last.stderr)}"
                 )
     check_default_exit()
+
+
+# --- Discovery & runner (CLI) ----------------------------------------------
+
+
+@dataclass
+class Result:
+    file: Path
+    name: str
+    status: str  # "pass" | "skip" | "fail"
+    message: str = ""
+
+
+def discover(root: Path) -> list[Path]:
+    return sorted(root.glob("docs/**/*.spec.md"))
+
+
+def run_file(path: Path, repo_root: Path) -> list[Result]:
+    return [run_test(test, path, repo_root) for test in parse_spec(path.read_text())]
+
+
+def run_test(test: SpecCase, path: Path, repo_root: Path) -> Result:
+    if test.skip:
+        return Result(path, test.name, "skip")
+    with tempfile.TemporaryDirectory() as tmp:
+        try:
+            execute(test, Path(tmp), repo_root)
+        except (SpecFailure, SpecError) as exc:
+            return Result(path, test.name, "fail", str(exc))
+    return Result(path, test.name, "pass")
+
+
+_GLYPH = {"pass": "✅", "skip": "🟡", "fail": "❌"}
+
+
+def main(argv: list[str] | None = None) -> int:
+    argv = sys.argv[1:] if argv is None else argv
+    root = Path.cwd()
+    files = sorted(Path(a) for a in argv) if argv else discover(root)
+    if not files:
+        print(f"no docs/**/*.spec.md under {root}")
+        return 1
+
+    counts = {"pass": 0, "skip": 0, "fail": 0}
+    for path in files:
+        rel = path.relative_to(root) if path.is_relative_to(root) else path
+        print(f"\n{rel}")
+        for r in run_file(path, root):
+            counts[r.status] += 1
+            print(f"  {_GLYPH[r.status]} {r.name}")
+            if r.status == "fail":
+                for line in r.message.splitlines():
+                    print(f"      {line}")
+
+    print(f"\n{counts['pass']} passed, {counts['skip']} skipped, {counts['fail']} failed")
+    return 1 if counts["fail"] else 0
+
+
+if __name__ == "__main__":
+    raise SystemExit(main())
